Update vendor Dashboard login color, including admin vendor header link

+++ app/Http/Controllers/StripeController.php
@@ -131,7 +131,9 @@ class StripeController extends Controller
         if(Session::has('coupon')){
             $total_amount = Session::get('coupon')['total_amount'];
         }else{
-            $total_amount = round(Cart::total());
+            
+            $total_amount = round((float)Cart::total());
+            
         }
